added contact page
contact page will take in your name, email, and a message and pass it through a submit event. Submit isn't fully done yet, the form will not do anything with this data yet. Each field will check if they are completed, and the email field will check for invalid emails.